fix: Add custom option to extract request information from context
Include a new parameter in the authentication guard that allows users to
customize how the request and response values are extracted from the
nestjs context.
This is required for people that use graphql.
We are using a custom method instead of defining different graphql
specific logic to prevent adding graphql specific libraries to our
dependencies.