fix: Mark the guard dependency as optional
The 0.0.3 change introduced a dependency regression in the
SuperTokensAuthGuard. The automated tests did not catch this because the
behavior doesn't seem to be reproductible in that environment. The
dependency error is triggered only when you run/build the NestJS
project. Hence, we should also confirm that the example apps work
properly with the latest changes.